requirements.txt
# Core dependencies
torch==2.8.0
torchvision==0.23.0
numpy==2.2.6
scipy==1.16.3

# Data handling
opencv-python==4.12.0.88
Pillow==12.0.0
h5py==3.15.1

# Visualization
matplotlib==3.10.7
seaborn==0.13.2

# Jupyter & notebooks
jupyter==1.1.1
ipython==9.8.0
ipywidgets==8.1.8

# Utilities
tqdm==4.67.1
pyyaml==6.0.3
tensorboard==2.20.0

# Object Detection
ultralytics==8.3.235

# 3D Geometry
trimesh==4.10.1

# Development
black==25.12.0
flake8==7.3.0
pytest==9.0.2